pyproject.toml
[tool.poetry]
name = "trainblock"
version = "0.0.1"
description = "Component building blocks for supervised ML training"
authors = ["loppjp <jacob.lopp@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
torch = "^2.8.0"
pyyaml = "^6.0.2"
opencv-python = ">=4.0,<5.0"
numpy = "^2.3.2"
tqdm = "^4.67.1"
torchvision = "^0.23.0"

[tool.poetry.dev-dependencies]
pytest = "^8.4.1"

[tool.poetry.extras]
formatting = ["black"]

[tool.poetry.scripts]
trainblock = 'trainblock.cli:main'

[tool.poetry.group.notebooks.dependencies]
matplotlib = "^3.10.5"
ipykernel = "^6.30.1"

[tool.poetry.group.dev.dependencies]
black = {extras = ["jupyter"], version = "^25.1.0"}
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"